+++ angular-research/src/app/features/standalone-components/custom-button/custom-button.component.ts
@@ -1,0 +1,17 @@
+import { Component, Input } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+@Component({
+  selector: 'custom-button',
+  standalone: true,
+  imports: [CommonModule],
+  templateUrl: './custom-button.component.html',
+  styleUrls: ['./custom-button.component.scss'],
+})
+export class CustomButtonComponent {
+  @Input('button-title') buttonTitle = '';
+
+  btnClickEvent() {
+    console.log('button click event');
+  }
+}