+++ angular-research/src/app/features/standalone-components/custom-input/custom-input.component.ts
@@ -1,0 +1,13 @@
+import { Component } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+@Component({
+  selector: 'custom-input',
+  standalone: true,
+  imports: [CommonModule],
+  templateUrl: './custom-input.component.html',
+  styleUrls: ['./custom-input.component.scss']
+})
+export class CustomInputComponent {
+
+}