+++ angular-research/ReleaseNotes.md
@@ -1,4 +1,7 @@
 Relase Notes:
 -------------
 
-* Test
+11/04/2023
+----------
+* Standalone components (custom-button)
+* Lazy standalone components(custom-input)